Proof of concept - full comment caching
OMG this is an awesome bit of python that I had no idea you could use before

This creates a function that stores the output of the function `get_code_commentors` for an entire repo call. This means that repeated calls do not have to re-get commentors.

Results are saved to file in a json format, as to not introduce yet-another-keyvalue-store.

This isn't as effective as it could be, because it stores the full result as a list of people. For proper effectiveness, results should be stored on a per pull/issue bases, as so to allow for incremental checks (e.g. 5000 tokens used for a repo with 5000+ issues will mean multiple runs spaced out. Storing the results of previous runs will build up a store of data, which on the final run will just complete 'instantly' rather than have to use any requests)

Thanks to @aurryn, @tveastman and @ncoghlan for their assistance :D