Prevent mutating cache
'Mutating cache': running octohatrack.py with a sufficiently full cache_file.json for a repo was resulting in an increating cache. This was reproducible even when the script was running on a machine lacking network.

The results for a pull/comments were increasing on subsequent runs. A result with 3 comments would increase to 6, then 9.

Fact: for a PR in github, content for comments appears at both the pulls/ID/comments and issues/ID/comments endpoints. The counters for these will be the same.

Speculation: the results for the issues were being appended to the results for the pulls. Hence a constant increase and not a multiplicative increase in the results.

I have no idea *how* it was mutating, but removing the `+=` of the results to loop through once, and processing the results of two memoise'd calls separately means the cache no longer mutates.